updated slider styles
